Warings y Version 1.3.4

+++ MonkeyDelivery/Src/Control/States/CreditsState.cpp
@@ -40,7 +40,7 @@ void CreditsState::update()
 {
 	if (index_ == credits_.size()) return;
 
-	if (SDL_GetTicks() > timeNewName_ + lastName_)
+	if ((int)SDL_GetTicks() > timeNewName_ + lastName_)
 	{
 		index_++;
 		lastName_ = SDL_GetTicks();
