+++ MonkeyDelivery/Src/Logic/DialogueBox.cpp
@@ -19,16 +19,16 @@ DialogueBox::DialogueBox(Game* game) : GameObject(game)
 
 
 	
-	yPos_ = game->getWindowHeight()/1.43;
-	width_ = game->getWindowWidth()/7.5 * 2;
-	xPos_ = game->getWindowWidth() / 2 - width_ / 2;
-	height_ = game->getWindowHeight()/11.37 * 2;
-	xText_ = game->getWindowWidth()/120.0;
-	yText_ = game->getWindowHeight()/50.0;
-	xIcon = - game->getWindowWidth()/36.0;
-	yIcon = -game->getWindowHeight()/17.24;
-	wIcon = game->getWindowWidth()/14.4;
-	hIcon = game->getWindowHeight()/8.0;
+	yPos_ = (int)(game->getWindowHeight()/1.43);
+	width_ = (int)(game->getWindowWidth()/7.5 * 2);
+	xPos_ = (int)(game->getWindowWidth() / 2 - width_ / 2);
+	height_ = (int)(game->getWindowHeight()/11.37 * 2);
+	xText_ = (int)(game->getWindowWidth()/120.0);
+	yText_ = (int)(game->getWindowHeight()/50.0);
+	xIcon = (int)(- game->getWindowWidth()/36.0);
+	yIcon = (int)(-game->getWindowHeight()/17.24);
+	wIcon = (int)(game->getWindowWidth()/14.4);
+	hIcon = (int)(game->getWindowHeight()/8.0);
 	
 	
 
@@ -66,7 +66,7 @@ void DialogueBox::changeMissionText(string id) {
 void DialogueBox::inShow()
 {
 	//actualizacion por tiempo
-	if (lastUpdate_ + showLetterTime_ > SDL_GetTicks())
+	if (lastUpdate_ + showLetterTime_ > (int)SDL_GetTicks())
 		return;
 
 	lastUpdate_ = SDL_GetTicks();
